Development

+++ backend/controllers/allergens.js
@@ -1,16 +1,21 @@
 import Allergen from '../models/Allergen.js'
+import createError from '../utils/error.js';
 
 export const createAllergen = async (req, res, next) => {
 
-    req.body.lastManipulatorId = req.client.id;
-    const newAllergen = new Allergen(req.body);
-
     try {
-        const savedAllergen = await newAllergen.save();
-        res.status(200).json(savedAllergen);
-        console.log(`${savedAllergen.allergenName} - allergen has been saved!`);
+        if (req.body.allergenName !== undefined && req.body.allergenName !== "") {
+            req.body.lastManipulatorId = req.client.id;
+            const newAllergen = new Allergen(req.body);
+            const savedAllergen = await newAllergen.save();
+            res.status(200).json(savedAllergen);
+            console.log(`${savedAllergen.allergenName} - allergen has been saved!`);
+        } else {
+            return next(createError(403, "Wrong data (allergen name is wrong or missing!"));
+        }
+    }
 
-    } catch (error) {
+    catch (error) {
         next(error);
     }
 };
@@ -18,10 +23,22 @@ export const createAllergen = async (req, res, next) => {
 export const getAllergens = async (req, res, next) => {
 
     try {
-        const allergens = (await Allergen.find()).filter((data) => data.isActive);
-        res.status(200).json(allergens);
+        let allergens = null;
+        if (req.query.isActive === 'true') {
+            allergens = (await Allergen.find()).filter((data) => data.isActive === true);
+        } else if (req.query.isActive === 'false') {
+            allergens = (await Allergen.find()).filter((data) => data.isActive === false);
+        } else {
+            allergens = await Allergen.find();
+        }
+        if (allergens !== null) {
+            res.status(200).json(allergens);
+        } else {
+            res.status(200).json("No allergens in the database!");
+        }
+    }
 
-    } catch (error) {
+    catch (error) {
         next(error);
     }
 };
@@ -30,7 +47,11 @@ export const getAllergenById = async (req, res, next) => {
 
     try {
         const currentAllergen = await Allergen.findById(req.params.id);
-        res.status(200).json(currentAllergen);
+        if (currentAllergen !== null) {
+            res.status(200).json(currentAllergen);
+        } else {
+            res.status(200).json("No allergen the given ID!");
+        }
 
     } catch (error) {
         next(error);
@@ -39,17 +60,28 @@ export const getAllergenById = async (req, res, next) => {
 
 export const updateAllergenById = async (req, res, next) => {
 
-    req.body.lastManipulatorId = req.client.id;
-
     try {
-        const updateAllergen = await Allergen.findByIdAndUpdate(
-            req.params.id,
-            { $set: req.body },
-            { new: true });
-        res.status(200).json(updateAllergen);
-        console.log(`${updateAllergen.allergenName} - ${updateAllergen._id} - has been updated!`);
+        const actualAllergen = await Allergen.findById(req.params.id);
+        if (actualAllergen !== null) {
 
-    } catch (error) {
+            if (req.body.allergenName !== undefined && req.body.allergenName !== "") {
+                req.body.lastManipulatorId = req.client.id;
+                const updateAllergen = await Allergen.findByIdAndUpdate(
+                    req.params.id,
+                    { $set: req.body },
+                    { new: true });
+
+                res.status(200).json(updateAllergen);
+                console.log(`${updateAllergen.allergenName} - ${updateAllergen._id} - has been updated!`);
+            } else {
+                return next(createError(403, "Wrong data (allergen name is wrong or missing!"));
+            }
+        } else {
+            res.status(200).json("No allergen the given ID!");
+        }
+    }
+
+    catch (error) {
         next(error);
     }
 };
@@ -57,11 +89,17 @@ export const updateAllergenById = async (req, res, next) => {
 export const deleteAllergenById = async (req, res, next) => {
 
     try {
-        await Allergen.findByIdAndDelete(req.params.id);
-        res.status(200).json(`${req.params.id} - allergen has been deleted!`);
-        console.log(`${req.params.id} - allergen has been deleted!`);
+        const actualAllergen = await Allergen.findById(req.params.id);
+        if (actualAllergen !== null) {
+            const deletedAllergen = await Allergen.findByIdAndDelete(req.params.id);
+            res.status(200).json(`${deletedAllergen._id} - allergen has been deleted!`);
+            console.log(`${deletedAllergen._id} - allergen has been deleted!`);
+        } else {
+            res.status(200).json("No allergen the given ID!");
+        }
+    }
 
-    } catch (error) {
+    catch (error) {
         next(error);
     }
 };
